Sagar_Dataharvesting/GDFL_upload/TradingDaysValidator/requirements.txt
fastapi==0.114.0
mysql_connector_repackaged==0.3.1
pydantic==2.9.1
SQLAlchemy==2.0.25
